5. zadatak za 4. cas prebacio da je long niz a ne int

+++ cas04/5_obrni/obrni.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
-void obrni(int[], int) ;
+void obrni(long[], int) ;
 
 int main()
 {
-	int a[] = {1,2,3,4,5,6,7,8};
+	long a[] = {1,2,3,4,5,6,7,8};
 	int n = 8, i;
 	printf("Pre obrtanja: ");
 	for(i=0; i<n; i++)
 	{
-		printf("%d ", a[i]);
+		printf("%ld ", a[i]);
 	}
 	printf("\n");
 	
@@ -18,7 +18,7 @@ int main()
 	printf("Posle obrtanja: ");
 	for(i=0; i<n; i++)
 	{
-		printf("%d ", a[i]);
+		printf("%ld ", a[i]);
 	}
 	printf("\n");
